Defer pypylon update
Defer pypylon 4.2.0 update; on macOS, we now fail to recognize two
of bundled shared libraries as dependencies and therefore do not
symlink them to the top-level application directory. While we
could try to work around this in several ways, the proper fix is
probably to fix the binary dependency analysis on macOS to recursively
analyze each passed binary (similarly to `ldd` on other POSIX
platforms), which would allow us to properly account for inheritance
of RPATHs from the referring binaries.