Adjust the automatically-raised recursion limit value for python < 3.11 on Windows
The recursion limit test in the main PyInstaller repository seems to
crash the python process when ran under Windows or Cygwin with
python < 3.11 and recursion limit is set to > 2097.

Therefore, for the affected combinations of platform and python version,
adjust the automatically-raised recursion limit value to a conservative
value of 2000. For other combinations, continue using the value of 5000.